refactor: rename some files and small fix
