+++ src/app/formly-config.module.ts
@@ -2,9 +2,9 @@ import { NgModule, ModuleWithProviders, Provider } from '@angular/core';
 import { SharedModule } from './shared/shared.module';
 
 import { FormlyModule } from '@ngx-formly/core';
-import { FormlyFieldComboboxComponent } from './formly-template';
-import { FormlyWrapperPanelComponent } from './formly-wrapper';
-import { FormlyValidation } from './formly-validation';
+import { FormlyFieldComboboxComponent } from './formly-templates';
+import { FormlyWrapperPanelComponent } from './formly-wrappers';
+import { FormlyValidations } from './formly-validations';
 
 /**
  * Formly global configuration
@@ -18,11 +18,11 @@ const formlyModuleProviders = FormlyModule.forRoot({
 @NgModule({
   imports: [SharedModule],
   declarations: [FormlyFieldComboboxComponent, FormlyWrapperPanelComponent],
-  providers: [FormlyValidation],
+  providers: [FormlyValidations],
 })
 export class FormlyConfigModule {
-  constructor(formlyValidation: FormlyValidation) {
-    formlyValidation.init();
+  constructor(formlyValidations: FormlyValidations) {
+    formlyValidations.init();
   }
 
   static forRoot(): ModuleWithProviders<FormlyConfigModule> {
